+++ src/thunderbots/software/ai/primitive/direct_wheels_primitive.cpp
@@ -1,5 +1,4 @@
-
-#include "ai/primitive/directwheels_primitive.h"
+#include "ai/primitive/direct_wheels_primitive.h"
 
 #include <cstdint>
