+++ src/thunderbots/software/ai/primitive/primitive_factory.cpp
@@ -9,7 +9,7 @@
 #include "ai/primitive/catch_primitive.h"
 #include "ai/primitive/chip_primitive.h"
 #include "ai/primitive/direct_velocity_primitive.h"
-#include "ai/primitive/directwheels_primitive.h"
+#include "ai/primitive/direct_wheels_primitive.h"
 #include "ai/primitive/dribble_primitive.h"
 #include "ai/primitive/kick_primitive.h"
 #include "ai/primitive/move_primitive.h"
